+++ PluginInstallation/CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [3.3.1](https://github.com/packlink-dev/prestashop_module/compare/v3.3.0...v3.3.1)
+### Changed
+- Add security improvements
+- Updated to the latest Core version 3.5.1
+
+## [v3.3.0](https://github.com/packlink-dev/prestashop_module/compare/v3.2.23...v3.3.0) - 2025-03-27
+### Added
+- Support for the special services
+- Manual refresh for shipping services
+### Changed
+- Updated to the latest Core version 3.5.0
+- 
 ## [3.2.23](https://github.com/packlink-dev/prestashop_module/compare/v3.2.22...v3.2.23)
 ### Changed
 - Restored shop context and Shop ID in install() to fix ps_mbo incompatibility.